merged mock fixes with navbar update (#7)
Fix build errors. This was due to a Node module configuration change to browser context. Functions like promisify or fs.readFile and not available in a browser context. The fix uses native typescript import to pull in the JSon, and no longer performs file reads. This mocking solution will work in both node and browser contexts.